Added support for IPS patches.
It's possible to patch the ROM through 3 ways:
1) rename the patch with the same name as the ROM and copy it to the
same directory where the ROM is present (example: if "Adventures of Lolo
(U) [!].nes" is the name of the ROM, rename the patch in "Adventures of
Lolo (U) [!].ips")
2) drag it to the emulator window (compressed patches are supported)
3) apply it via the "Apply IPS Patch" item in the File menu (compressed
patches are supported)
If more than one patch is present in the compressed file, it will always
be possible to select which one to apply.
In order to automatically apply a patch to a ROM, in addition to method
1) it is also possible to create a compressed archive that contains both
the ROM and the patch. In this case the name of the ips file isn't
important (example : comressed file contain "Adventures of Lolo (U)
[!].nes" and "adv.ips").