Added the rewind function that replaces the timeline.
Now it is possible to "rewind" the game. Once the rewind function has
been activated (at any time during the game) it is possible to move
forward and backward in the gameplay one frame at a time or at different
speeds (1x, 2x, 4x, 8x, 16x, 32x or 64x). You will be able to move
freely in the recording until you press the "play" button, from that
moment all the gameplay memorized after the selected frame will be lost.
In the "general settings" you can select the duration in minutes of the
recordable gameplay (the different options are: disable, 2mins, 5mins,
15mins, 30min, 60mins or unlimited) but keep in mind that the more
minutes you can record the more space on the hard disk will be used.