Added the preview of the last screen image of the saved states.
Now it is always possible view the screen of the saved states when you
move the mouse between the various Slots in the "State" menu (slots with
the available preview will be marked with a "*"). It is also available
in the combobox of the toolbar widget.The preview will be displayed as a
tooltip.