Implemented the ability to select the resolution of the fulscreen.
Also added the option to automatically set the refresh rate based on the
region of the rom used (50/100Hz for PAL/Dendy roms, 60/120Hz for NTSC)
if the monitor supports them. The rewriting of the fullscreen management
allowed to correct some bugs present for a long time (for example under
Linux/BSD the fullscreen window did not always identify the real screen
size available because it did not take into account the various task
bars and menu bars).