Added a new tool: "Joypad/Gamepad Debug".
With this you can see the info on the connected devices, it is also
possible to disable the reading of certain axes and buttons from the
gamepad status reading routine. In addition to debugging reasons, it is
an option to be able to use gamepads that send incorrect values on some
axis or button while still allowing their use.