I implemented the "Paste" function for the Family Basic Keyboard.
It's possible to paste the text by dragging it in the main window of the
emulator or copying it and then press the right mouse button and
selecting "Paste". The text will be inserted through the simulation of
the pressure of the keys and it will always be possible to interrupt the
insertion by pressing the ESC key. During the paste, the Fast Forwandig
will be active to reduce the waiting time. It was not possible for me to
simulate the pressure of the Japanese characters because I don't know
the language, if someone wants to help me can contact me.