Support for XDG Base Directory Specification.
On UNIX-Like operating systems the configuration files will be found in
"$XDG_CONFIG_HOME/puNES" and the rest in "$XDG_DATA_HOME/puNES".
Also under Windows the location of the working directory has changed and
will now be located in "C:/Users/<USER>/AppData/Local/puNES".
Regarding the portable mode nothing has changed, everything will be
placed in the directory containing the executable.
When the emulator is started, the old directories will be automatically
moved to the new location.
In order to open the configuration directory easily, it's always
possible to do so via the menu items "File->Open working directory" or
"File->Open config directory".